changed spreadsheet parse output to use a HashSet instead of a new SpreadsheetData type.
Created new SpreadsheetData record to hold more than just two parameters of data

This allows the passing of extra data besides the name and guess columns, and can now include the row number. This can be used to provide more helpful feedback to user.

The HashSet type gives the benefits of a Dictionary in that there is a simple check if duplicate add being done.